pyproject.toml
[tool.poetry]
name = "testtask"
version = "0.1.0"
description = ""
authors = ["LonerLykos <loner.lykos.project@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "3.16.0"
mysqlclient = "^2.2.7"
django-filter = "^25.1"
djangorestframework-simplejwt = "^5.5.0"
pillow = "^11.1.0"

[tool.poetry.group.dev.dependencies]
isort = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"